add support for sampling_ratio
sampling_ratio allows users to specify if they want to drop certain mappings in the chunk: qna pairing block.
The amount of these mappings per chunk is quite high for users running the full pipeline on a laptop.

While it is an issue that SDG itself takes a while the larger issue is that training the sometimes very large dataset produced by a 24hr+ SDG run is sometimes not feasible on a consumer laptop.

This knob is meant to be something auto-detected as part of system profiles. A user running on an M Series Mac should not expect 500+ generated samples taking 24 hrs. More realistically, they want ~100 samples maximum in ~2 hours. In testing, this sampling ratio cuts the generation time significantly!

Signed-off-by: Charlie Doern <cdoern@redhat.com>